Rename headers in src/external/metis4/include to avoid conflicts
