+++ src/external/metis4/test_nodend2.c
@@ -13,7 +13,7 @@
  *
  */
 
-#include "metis.h"
+#include "metis4.h"
 #include <time.h>
 
 int main( )
