+++ src/external/metis4/util.c
@@ -17,7 +17,7 @@
  *
  */
 
-#include "metis.h"
+#include "metis4.h"
 
 /* GKlib utilities */
